Refactor Software component to create a new Portfolio layout with navigation and hero sections
